+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# [1.2.4-rc.4]
+
+- (QA) Fixed unable to hover over aircraft issue
+- (QA) Removed Pizza Hut from detailed map layer
+- You can now choose to show raw SIGMET data only when clicking on them
+- Default ground traffic mode is now "hide if zoomed out" instead of "show all"
+
 # [1.2.4-rc.3]
 
 - NY_CTR has finally been excluded from VATGlasses display